--- mythril/ether/__init__.py
@@ -1,3 +1,0 @@
-import time
-
-start_time = time.time()
