+++ mythril/ethereum/util.py
