+++ tests/ethcontract_test.py
@@ -1,5 +1,5 @@
 import unittest
-from mythril.ether.ethcontract import ETHContract
+from mythril.ethereum.ethcontract import ETHContract
 
 
 class ETHContractTestCase(unittest.TestCase):
