+++ tests/laser/transaction/create_transaction_test.py
@@ -1,9 +1,9 @@
 from mythril.laser.ethereum.transaction import execute_contract_creation
-from mythril.ether import util
+from mythril.ethereum import util
 import mythril.laser.ethereum.svm as svm
 from mythril.disassembler.disassembly import Disassembly
 from datetime import datetime
-from mythril.ether.soliditycontract import SolidityContract
+from mythril.solidity.soliditycontract import SolidityContract
 import tests
 from mythril.analysis.security import fire_lasers
 from mythril.analysis.symbolic import SymExecWrapper
