+++ tests/report_test.py
@@ -1,10 +1,9 @@
 from mythril.analysis.report import Report
 from mythril.analysis.security import fire_lasers
 from mythril.analysis.symbolic import SymExecWrapper
-from mythril.ether import util
-from mythril.ether.soliditycontract import ETHContract
+from mythril.ethereum import util
+from mythril.solidity.soliditycontract import ETHContract
 from multiprocessing import Pool, cpu_count
-import datetime
 import pytest
 import json
 from tests import *
